requirements.txt
pandas>=2.0,<3.0
numpy>=1.24,<2.0
scipy>=1.11,<2.0
pyyaml>=6.0,<7.0
openpyxl>=3.1,<4.0
plotly>=5.18,<6.0
